// apps/sage-sync/src/lib/sage/types.ts
type SageRecord = {
  RECORD_MODIFY_DATE: string;
  RECORD_CREATE_DATE: string;
  RECORD_DELETED: 0 | 1;
};

export type DEPARTMENT = SageRecord & {
  NUMBER: number;
  NAME: string;
  REFERENCE: string;
};

export type STOCK_CAT = {
  NUMBER: number;
  NAME: string;
};

export type PURCHASE_LEDGER = SageRecord & {
  ACCOUNT_REF: string;
  NAME: string;
};

export type PURCHASE_ORDER = SageRecord & {
  ORDER_NUMBER: number;
  ACCOUNT_REF: string;
  ORDER_DATE: string;
  ORDER_OR_QUOTE: string;
  ORDER_STATUS: string;
  ORDER_STATUS_CODE: number;
  DELIVERY_DATE: string;
  DELIVERY_STATUS: string;
  DELIVERY_STATUS_CODE: number;
};

export type POP_ITEM = SageRecord & {
  ITEMID: number;
  ORDER_NUMBER: number;
  STOCK_CODE: string;
  DELIVERY_DATE: string;
  DUE_DATE: string | null;
  QTY_ALLOCATED: number;
  QTY_DELIVERED: number;
  QTY_DESPATCH: number;
  QTY_ORDER: number;
};

export type SALES_LEDGER = SageRecord & {
  ACCOUNT_REF: string;
  NAME: string;
};

export type SALES_ORDER = SageRecord & {
  ORDER_NUMBER: number;
  ACCOUNT_REF: string;
  ORDER_DATE: string;
  ORDER_OR_QUOTE: string;
  ORDER_TYPE_CODE: number;
  DESPATCH_DATE: string | null;
  DESPATCH_STATUS: string;
  DESPATCH_STATUS_CODE: number;
};

export type SOP_ITEM = SageRecord & {
  ITEMID: number;
  ORDER_NUMBER: number;
  STOCK_CODE: string;
  DELIVERY_DATE: string;
  DUE_DATE: string | null;
  QTY_ALLOCATED: number;
  QTY_DELIVERED: number;
  QTY_DESPATCH: number;
  QTY_ORDER: number;
};
// ...
export type STOCK_TRAN = SageRecord & {
  TRAN_NUMBER: number;
  STOCK_CODE: string;
  TYPE: string;
  DATE: string;
  REFERENCE: string;
  REFERENCE_NUMERIC: number;
  DETAILS: string;
  QUANTITY: number;
  COST_PRICE: number;
  SALES_PRICE: number;
  QTY_USED: number;
  ISP_REFERENCE: number;
  ARTEFACT_TYPE: number;
  GDN_NUMBER: number;
  GRN_NUMBER: number;
};

export type GDN_ITEM = SageRecord & {
  UNIQUE_ID: number;
  GDN_NUMBER: number;
  ITEM_NUMBER: number;
  ORDER_NUMBER: number;
  ORDER_ITEM: number;
  ACCOUNT_REF: string;
  CUSTOMER_GDN_NUMBER: string;
  STOCK_CODE: string;
  DESCRIPTION: string;
  DATE: string;
  QTY_DESPATCHED: number;
  QTY_ON_ORDER: number;
  PRINTED_FLAG: number;
};

export type GRN_ITEM = SageRecord & {
  GRN_NUMBER: number;
  ITEM_NUMBER: number;
  ORDER_NUMBER: number;
  ORDER_ITEM: number;
  ACCOUNT_REF: string;
  SUPPLIER_GRN_NUMBER: string;
  STOCK_CODE: string;
  DESCRIPTION: string;
  DATE: string;
  QTY_RECEIVED: number;
  QTY_ON_ORDER: number;
};